+++ src/indexable/dataloader/mod.rs
@@ -46,7 +46,7 @@ where
     D: Dataset,
     DefaultCollate: Collate<D::Sample>,
 {
-    /// Helper to return a [`Builder`] easily.
+    /// Helper to return a [`DataLoader`] builder.
     pub fn builder(dataset: D) -> Builder<D, SequentialSampler, DefaultCollate> {
         Builder::new(dataset)
     }