+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+## [0.3.0] - 2022-01-11
 ### Added
 - more strict compiler warnings
 - `Dataloader` for iterable datasets. A dalaloader working for any type that implements `IntoIterator`
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Sequential and random `Sampler`
 
 
-[Unreleased]: https://github.com/Tudyx/ai-dataloader/compare/v0.2.1...HEAD
+[Unreleased]: https://github.com/Tudyx/ai-dataloader/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/Tudyx/ai-dataloader/compare/v0.2.1...v0.3.0
 [0.2.1]: https://github.com/Tudyx/ai-dataloader/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/Tudyx/ai-dataloader/compare/v0.1.0...v0.2.0